+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix muting bug with FXH parser [@coffeevalenbat](https://github.com/coffeevalenbat)
 - Fix issue where Animation State value for projectiles was being ignored
 - Fix issue where falling on to ladder while holding dpad down could sometimes cause player to get stuck
+- Fix bounds check for right screen edge when player isn't 16px wide
 
 ### Performance